Define the loadModule events in a separate file.
- This allows both the client & the server to depend on it sanely.
- Still passes deno check
- Makes vscode happy.
